+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # v1.3.0
 * V12 compatibility
-* Add option to remove GM hidden state on combatants at combat start.
+* Add option to remove GM hidden state on noticed combatants at combat start.
 
 # v1.2.3
 * Fix failed stealth to follow Sneak rules